MySQL driver builds

+++ ORDA/Implementation/ORDAStatementResultImpl.h
@@ -15,7 +15,7 @@
  */
 @interface ORDAStatementResultImpl : ORDAResultImpl <ORDAStatementResult>
 
-@property (readonly) int changed;
+@property (readonly) long long changed;
 @property (readonly) long long lastID;
 @property (readonly) NSDictionary * dict;
 @property (readonly) NSArray * array;
@@ -23,7 +23,7 @@
 //+ (NSDictionary *)arrayDictFromDictArray:(NSArray *)array andRows:(int)rows andColumns:(NSArray *)columns;
 //+ (NSArray *)dictArrayFromArrayDict:(NSDictionary *)dict andRows:(int)rows andColumns:(NSArray *)columns;
 
-+ (ORDAStatementResultImpl *)statementResultWithChanged:(int)changed andLastID:(long long)lastID andRows:(int)rows andColumns:(NSArray *)columns andDictionaryOfArrays:(NSDictionary *)dict andArrayOfDictionaries:(NSArray *)array;
-- (id)initWithChanged:(int)changed andLastID:(long long)lastID andRows:(int)rows andColumns:(NSArray *)columns andDictionaryOfArrays:(NSDictionary *)dict andArrayOfDictionaries:(NSArray *)array;
++ (ORDAStatementResultImpl *)statementResultWithChanged:(long long)changed andLastID:(long long)lastID andRows:(long)rows andColumns:(NSArray *)columns andDictionaryOfArrays:(NSDictionary *)dict andArrayOfDictionaries:(NSArray *)array;
+- (id)initWithChanged:(long long)changed andLastID:(long long)lastID andRows:(long)rows andColumns:(NSArray *)columns andDictionaryOfArrays:(NSDictionary *)dict andArrayOfDictionaries:(NSArray *)array;
 
 @end
